+++ interfaces/robot_motion_interface/robot_motion_interface_py/src/robot_motion_interface/panda/panda_interface.py
@@ -1,20 +1,15 @@
-from robot_motion_interface_py.interface import Interface
-from robot_motion_interface_py.panda.panda_interface import PandaInterface
-from robot_motion_interface_py.tesollo.tesollo_interface import TesolloInterface
- 
+from robot_motion_interface.interface import Interface
 from enum import Enum
 import numpy as np
 
 
-class PandaTesolloUnifiedInterface(Interface):
+class PandaInterface(Interface):
     
     def __init__(self):
         """
-        Wrapper for using Tesollo attached to Panda (mainly made for unified IK) 
+        Python wrapper for C++ Panda Interface.
         """
-        self.panda = PandaInterface()
-        self.tesollo = TesolloInterface()
-
+        self._start_loop()
 
 
     def set_joint_positions(self, q:np.ndarray, joint_names:list[str] = None, blocking:bool = False):
@@ -28,7 +23,6 @@ class PandaTesolloUnifiedInterface(Interface):
             blocking (bool): If True, the call should returns only after the controller
                 achieves the target. If False, returns after queuing the request.
         """
-        # Todo send panda joints to panda interface and tesollo joints to tesollo interface
         ...
     
     def set_cartesian_pose(self, x:np.ndarray,  base_frame:str = None, ee_frames:list[str] = None, blocking:bool = False):
@@ -46,9 +40,16 @@ class PandaTesolloUnifiedInterface(Interface):
             blocking (bool): If True, the call returns only after the controller
                 achieves the target. If False, returns after queuing the request.
         """
-        # TODO: calculate IK together and call set_joint_position both
         ...
 
+    def set_control_mode(self, control_mode: Enum):
+        """
+        Set the control mode.
+
+        Args:
+            control_mode (Enum): Desired mode.Exact options are implementation-specific.
+        """
+        ...
     
     def home(self, blocking:bool = True):
         """
@@ -58,8 +59,7 @@ class PandaTesolloUnifiedInterface(Interface):
             blocking (bool): If True, the call returns only after the controller
                 homes. If False, returns after queuing the home request.
         """
-        self.tesollo.home(blocking=False)
-        self.panda.home(blocking=blocking)
+        ...
     
 
     def joint_positions(self) -> np.ndarray:
@@ -69,8 +69,7 @@ class PandaTesolloUnifiedInterface(Interface):
         Returns:
             (np.ndarray): (n_joints,) Current joint angles in radians.
         """
-        return np.concatenate([self.panda.joint_positions(), self.tesollo.joint_positions()])
-        
+        ...
 
     def joint_velocities(self) -> np.ndarray:
         """
@@ -79,9 +78,7 @@ class PandaTesolloUnifiedInterface(Interface):
         Returns:
             (np.ndarray): (n_joints,) Current joint velocities in radians.
         """
-        
-        return np.concatenate([self.panda.joint_velocities(), self.tesollo.joint_velocities()])
-        
+        ...
 
 
     def cartesian_pose(self, base_frame:str = None, ee_frame:str = None) -> np.ndarray:
@@ -95,8 +92,7 @@ class PandaTesolloUnifiedInterface(Interface):
             (np.ndarray): (7) Current pose in base frame [x, y, z, qx, qy, qz, qw]. 
                           Positions in m, angles in rad.
         """
-
-        return self.tesollo.cartesian_pose(base_frame, ee_frame)
+        ...
         
     
     def joint_names(self) -> list[str]:
@@ -106,9 +102,37 @@ class PandaTesolloUnifiedInterface(Interface):
         Returns:
             (list[str]): (n_joints) Names of joints
         """
-        return self.panda.joint_names() + self.tesollo.joint_names()
+        ...
     
 
+    ########################## Private ##########################
+    def _write_joint_torques(self, tau:np.ndarray):
+        """
+        Writes torque commands directly to motor.
+
+        Args:
+            tau (np.ndarray): (n_joints,) Commanded joint torques [N·m].
+        """
+        ...
+    
+
+    def _write_joint_positions(self, q:np.ndarray):
+        """
+        Write position commands directly to motor.
+
+        Args:
+            q (np.ndarray):  (n_joints,) Commanded joint positions [rad].
+        """
+        ...
+
+
+    def _start_loop(self):
+        """
+        Start the background runtime (e.g. for control loop and/or simulation loop).
+        """
+        ...
+
+
 if __name__ == "__main__":
-    panda = PandaTesolloUnifiedInterface()
+    panda = PandaInterface()
