+++ interfaces/robot_motion_interface/robot_motion_interface_py/src/robot_motion_interface/tesollo/tesollo_interface.py
@@ -1,4 +1,4 @@
-from robot_motion_interface_py.interface import Interface
+from robot_motion_interface.interface import Interface
 from enum import Enum
 import numpy as np
