+++ interfaces/sensor_interface/sensor_interface_py/src/sensor_interface/camera/realsense_interface.py
@@ -1,10 +1,8 @@
-from interfaces.sensor_interface.sensor_interface_py.src.sensor_interface_py.camera.rgbd_camera import CameraIntrinsics, RGBDCameraInterface, RGBDFrame
+from sensor_interface.camera.rgbd_camera import CameraIntrinsics, RGBDCameraInterface, RGBDFrame
 from typing import Literal
 import numpy as np
 
-
-
-class KinectInterface(RGBDCameraInterface):
+class RealsenseInterface(RGBDCameraInterface):
     def __init__(self, color_intrinsics: CameraIntrinsics, depth_intrinsics: CameraIntrinsics, T_color_depth:np.ndarray):
         """
         Initialize RGB-D interface.
@@ -22,7 +20,6 @@ class KinectInterface(RGBDCameraInterface):
             - Depth images are float32 meters.
         """
         super().__init__(color_intrinsics, depth_intrinsics, T_color_depth) 
-
 
 
     def start(self, resolution: tuple[int, int] = (640, 480), fps: int = 30,
@@ -74,7 +71,6 @@ class KinectInterface(RGBDCameraInterface):
 
 
 if __name__ == "__main__":
-   
     # MADE UP FOR TESTING
     color_intrinsics = CameraIntrinsics(
         width=640, height=480,
@@ -99,4 +95,4 @@ if __name__ == "__main__":
     ], dtype=np.float32)
     T_color_depth[:3,  3] = np.array([0.020, -0.010, 0.000], dtype=np.float32)  # meters
         
-    kinect = KinectInterface(color_intrinsics, depth_intrinsics, T_color_depth)
+    realsense = RealsenseInterface(color_intrinsics, depth_intrinsics, T_color_depth)
